extract importance and category from ToDoItem
