+++ ToDoList/ToDoList/Models/Importance.swift
@@ -1,0 +1,7 @@
+import Foundation
+
+enum Importance: String, Identifiable, Codable {
+    case low, basic, important
+    
+    var id: Self { self }
+}
